blawx/requirements.txt
djangorestframework
markdown
swiplserver
pyyaml
cobalt
clean-law >=0.0.4
django-guardian